requirements.txt
Django>=5.0
django-tenants>=3.5.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
mysqlclient>=2.2.0
requests>=2.31.0
pymysql>=1.1.0
